moved all images to /public folder
